+++ packages/flutter/lib/src/locales/fr.dart
@@ -26,6 +26,7 @@ const PlassLabels fr = PlassLabels(
   preview: 'Aperçu',
   empty: 'Rien ici',
   optional: 'Facultatif',
+  loading: 'Chargement',
   breadcrumb: 'Fil d’Ariane',
   breadcrumbExpand: 'Afficher les étapes masquées',
   carousel: 'Carrousel',