+++ packages/flutter/lib/src/locales/zh_hans.dart
@@ -26,6 +26,7 @@ const PlassLabels zhHans = PlassLabels(
   preview: '预览',
   empty: '暂无内容',
   optional: '可选',
+  loading: '加载中',
   breadcrumb: '面包屑导航',
   breadcrumbExpand: '显示折叠的层级',
   carousel: '轮播',