+++ packages/react/src/locales/es.ts
@@ -24,6 +24,7 @@ export const es: PlassLabels = {
   preview: 'Vista previa',
   empty: 'No hay nada aquí',
   optional: 'Opcional',
+  loading: 'Cargando',
 
   breadcrumb: 'Ruta de navegación',
   breadcrumbExpand: 'Mostrar los pasos ocultos',